requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
Deprecated==1.2.14
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
pycparser==2.22
PyGithub==2.3.0
Pygments==2.19.1
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
rich==14.0.0
tqdm==4.67.1
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
